MultiDataset class moved to datasets namespace
